backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sec-downloader
python-dotenv
lxml
yfinance
openai
google-genai
httpx
anyio
